-Modified BudgetManager.js to update a record into WalletTransaction table -Updated WalletTransactionManager.js to handle cases of adding balance or initiating refund -Integrated RazorPay APIs for adding wallet balance -Fixed a bug where sharerate was being parsed as int instead of float in ActivityTracker.js

+++ routes/cread/track/ActivityTracker.js
@@ -113,7 +113,7 @@ router.post('/', function (request, response) {
                                     .reduce(function (accumulator, element) {
                                         if (element.hasOwnProperty('sharerate')) {
                                             if (!element.donation) {
-                                                return accumulator + parseInt(element.sharerate);
+                                                return accumulator + parseFloat(element.sharerate);
                                             }
                                             else {
                                                 return accumulator;
